include error handling configuraiton
