--- edge-functions/packages/error-response/.env
@@ -1,1 +1,0 @@
-FRONTEND_HOST="somehost.com"
