--- edge-functions/packages/error-response/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@edge-functions/error-response",
-  "version": "1.0.0",
-  "description": "Error response Edge Functions",
-  "type": "module",
-  "main": "build/index.js",
-  "scripts": {
-    "build": "rm -rf build && node esbuild.js"
-  },
-  "dependencies": {
-    "commander": "^8.3.0",
-    "esbuild-runner": "^2.2.1"
-  }
-}
