--- edge-functions/packages/error-response/src/index.ts
@@ -1,56 +1,0 @@
-/**
- * This function is used to return the default page when a non-200 is returned from the origin.
- */
-
-import "source-map-support/register"
-import { CloudFrontRequest, CloudFrontResponseEvent, CloudFrontResponse } from "aws-lambda"
-import axios from "axios"
-import * as https from "https"
-
-const FRONTEND_HOST = process.env.FRONTEND_HOST
-const PATH_PREFIX = process.env.PATH_PREFIX || ""
-// Create axios client outside of lambda function for re-use between calls
-const instance = axios.create({
-  timeout: 1000,
-  // Don't follow redirects
-  maxRedirects: 0,
-  // Only valid response codes are 200
-  validateStatus: function (status) {
-    return status == 200
-  },
-  // keep connection alive so we don't constantly do SSL negotiation
-  httpsAgent: new https.Agent({ keepAlive: true }),
-})
-
-export const handler = (event: CloudFrontResponseEvent): Promise<CloudFrontResponse | CloudFrontRequest> => {
-  const response = event.Records[0].cf.response
-  const request = event.Records[0].cf.request
-
-  if (
-    response.status != "200" &&
-    !request.headers["x-request-prerender"] &&
-    request.uri != `${PATH_PREFIX}/index.html`
-  ) {
-    // Fetch default page and return body
-    return instance
-      .get(`https://${FRONTEND_HOST}${PATH_PREFIX}/index.html`)
-      .then(() => {
-        response.headers["content-type"] = [
-          {
-            key: "Content-Type",
-            value: "text/html",
-          },
-        ]
-
-        // Remove content-length if set as this may be the value from the origin.
-        delete response.headers["content-length"]
-
-        return response
-      })
-      .catch(() => {
-        return response
-      })
-  }
-
-  return Promise.resolve(response)
-}
