--- edge-functions/packages/error-response/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./build/"
-  },
-  "include": ["./**/*.ts"]
-}
